% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{ds_voucher_upload}
\alias{ds_voucher_upload}
\title{凭证数据源上传}
\usage{
ds_voucher_upload(
  dms_token,
  file_name = "data-raw/模板表上海嘉好20230715.xlsx"
)
}
\arguments{
\item{dms_token}{口令}

\item{file_name}{文件名}
}
\value{
返回值
}
\description{
凭证数据源上传
}
\examples{
ds_voucher_upload()
}
