% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voucher.R
\name{middlevourchermain}
\alias{middlevourchermain}
\title{Title 获取当前会计期间数据，生成凭证过程表}
\usage{
middlevourchermain(FToken, FYear, FMonth, FOpthon)
}
\arguments{
\item{FToken}{口令}

\item{FYear}{会计年度}

\item{FMonth}{会计期间}

\item{FOpthon}{权限}
}
\value{
是否运行成功
}
\description{
Title 获取当前会计期间数据，生成凭证过程表
}
\examples{
outputvourchermain("057A7F0E-F187-4975-8873-AF71666429AB", "2023", "4", '账套查询DMS测试')
}
