% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleTable.R
\name{odsmiddleTable}
\alias{odsmiddleTable}
\title{ods清除中间表}
\usage{
odsmiddleTable(token, FYear = "2023", FMonth = "8", FExpenseOrgID = "江苏")
}
\arguments{
\item{FExpenseOrgID}{}
}
\value{
无返回值
}
\description{
ods清除中间表
}
\examples{
odsmiddleTable()
}
