% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salary.R
\name{odssalary}
\alias{odssalary}
\title{ods工资}
\usage{
odssalary(token, FYear = "2023", FMonth = "8", FExpenseOrgID = "江苏")
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
ods工资
}
\examples{
odssalary()
}
