% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonrddetail.R
\name{srcnonrddetail}
\alias{srcnonrddetail}
\title{src工资}
\usage{
srcnonrddetail(token, FYear = "2023", FMonth = "8", FExpenseOrgID = "江苏")
}
\arguments{
\item{FExpenseOrgID}{}
}
\value{
无返回值
}
\description{
src工资
}
\examples{
srcnonrddetail()
}
