% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddetail.R
\name{srcrddetail}
\alias{srcrddetail}
\title{src工资}
\usage{
srcrddetail(token, FYear = "2023", FMonth = "8", FExpenseOrgID = "江苏")
}
\arguments{
\item{FExpenseOrgID}{}
}
\value{
无返回值
}
\description{
src工资
}
\examples{
srcrddetail()
}
