% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salary.R
\name{stdsalary}
\alias{stdsalary}
\title{std工资}
\usage{
stdsalary(token, FYear = "2023", FMonth = "8", FExpenseOrgID = "江苏")
}
\arguments{
\item{FExpenseOrgID}{}
}
\value{
无返回值
}
\description{
std工资
}
\examples{
stdsalary()
}
