% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialsecurity.R
\name{stdsocialsecurity}
\alias{stdsocialsecurity}
\title{std工资}
\usage{
stdsocialsecurity(token, FYear = "2023", FMonth = "8", FExpenseOrgID = "江苏")
}
\arguments{
\item{FExpenseOrgID}{}
}
\value{
无返回值
}
\description{
std工资
}
\examples{
stdsocialsecurity()
}
